+++ SDK/Extras/Textify3DMF/Source/Fetch.h
@@ -9,13 +9,13 @@
 
 void	SetBigEndian( bool inBigEndian );
 
-uint16_t FetchUShort( const uint8_t* inData, int32_t inOffset );
+uint16_t FetchUShort( const uint8_t* inData, size_t inOffset );
 
-uint32_t FetchULong( const uint8_t* inData, int32_t inOffset );
+uint32_t FetchULong( const uint8_t* inData, size_t inOffset );
 
-uint64_t FetchU64( const uint8_t* inData, int32_t inOffset );
+uint64_t FetchU64( const uint8_t* inData, size_t inOffset );
 
-float	FetchFloat( const uint8_t* inData, int32_t inOffset );
+float	FetchFloat( const uint8_t* inData, size_t inOffset );
 
-int32_t FetchLong( const uint8_t* inData, int32_t inOffset );
+int32_t FetchLong( const uint8_t* inData, size_t inOffset );
