+++ SDK/Extras/Textify3DMF/Source/TypeHandlers/Attributes/SurfaceTangent.cpp
@@ -32,8 +32,8 @@ SurfaceTangent::SurfaceTangent()
 {
 }
 
-void	SurfaceTangent::Process( uint32_t inStartOffset,
-									uint32_t inEndOffset )
+void	SurfaceTangent::Process( size_t inStartOffset,
+									size_t inEndOffset )
 {
 	if (inEndOffset != inStartOffset + 24)
 	{
