+++ SDK/Extras/Textify3DMF/Source/TypeHandlers/Attributes/TransparencyColor.cpp
@@ -32,8 +32,8 @@ TransparencyColor::TransparencyColor()
 {
 }
 
-void	TransparencyColor::Process( uint32_t inStartOffset,
-									uint32_t inEndOffset )
+void	TransparencyColor::Process( size_t inStartOffset,
+									size_t inEndOffset )
 {
 	if (inEndOffset != inStartOffset + 12)
 	{
