+++ SDK/Extras/Textify3DMF/Source/TypeHandlers/Geometries/Caps.cpp
@@ -37,8 +37,8 @@ Caps::Caps()
 {
 }
 
-void	Caps::Process( uint32_t inStartOffset,
-										uint32_t inEndOffset )
+void	Caps::Process( size_t inStartOffset,
+										size_t inEndOffset )
 {
 	if (inEndOffset != inStartOffset + 4)
 	{
