+++ SDK/Extras/Textify3DMF/Source/TypeHandlers/Geometries/Cone.cpp
@@ -32,10 +32,10 @@ Cone::Cone()
 {
 }
 
-void	Cone::Process( uint32_t inStartOffset,
-										uint32_t inEndOffset )
+void	Cone::Process( size_t inStartOffset,
+										size_t inEndOffset )
 {
-	uint32_t dataLen = inEndOffset - inStartOffset;
+	size_t dataLen = inEndOffset - inStartOffset;
 	
 	if (dataLen == 0)
 	{
