+++ SDK/Extras/Textify3DMF/Source/TypeHandlers/Geometries/Cylinder.cpp
@@ -32,10 +32,10 @@ Cylinder::Cylinder()
 {
 }
 
-void	Cylinder::Process( uint32_t inStartOffset,
-										uint32_t inEndOffset )
+void	Cylinder::Process( size_t inStartOffset,
+										size_t inEndOffset )
 {
-	uint32_t dataLen = inEndOffset - inStartOffset;
+	size_t dataLen = inEndOffset - inStartOffset;
 	
 	if (dataLen == 0)
 	{
