+++ SDK/Extras/Textify3DMF/Source/TypeHandlers/Geometries/Ellipse.cpp
@@ -32,10 +32,10 @@ Ellipse::Ellipse()
 {
 }
 
-void	Ellipse::Process( uint32_t inStartOffset,
-										uint32_t inEndOffset )
+void	Ellipse::Process( size_t inStartOffset,
+										size_t inEndOffset )
 {
-	uint32_t dataLen = inEndOffset - inStartOffset;
+	size_t dataLen = inEndOffset - inStartOffset;
 	
 	if (dataLen == 0)
 	{
