+++ SDK/Extras/Textify3DMF/Source/TypeHandlers/Geometries/Ellipsoid.cpp
@@ -32,10 +32,10 @@ Ellipsoid::Ellipsoid()
 {
 }
 
-void	Ellipsoid::Process( uint32_t inStartOffset,
-										uint32_t inEndOffset )
+void	Ellipsoid::Process( size_t inStartOffset,
+										size_t inEndOffset )
 {
-	uint32_t dataLen = inEndOffset - inStartOffset;
+	size_t dataLen = inEndOffset - inStartOffset;
 	
 	if (dataLen == 0)
 	{
