+++ SDK/Extras/Textify3DMF/Source/TypeHandlers/Geometries/TriGrid.cpp
@@ -33,10 +33,10 @@ TriGrid::TriGrid()
 {
 }
 
-void	TriGrid::Process( uint32_t inStartOffset,
-										uint32_t inEndOffset )
+void	TriGrid::Process( size_t inStartOffset,
+										size_t inEndOffset )
 {
-	uint32_t dataLen = inEndOffset - inStartOffset;
+	size_t dataLen = inEndOffset - inStartOffset;
 
 	if (dataLen < 8)
 	{
