+++ SDK/Extras/Textify3DMF/Source/TypeHandlers/Geometries/Triangle.cpp
@@ -33,9 +33,9 @@ Triangle::Triangle()
 {
 }
 
-void	Triangle::Process( uint32_t inStartOffset, uint32_t inEndOffset )
+void	Triangle::Process( size_t inStartOffset, size_t inEndOffset )
 {
-	uint32_t dataLen = inEndOffset - inStartOffset;
+	size_t dataLen = inEndOffset - inStartOffset;
 
 	if (dataLen != 36)
 	{
