+++ SDK/Extras/Textify3DMF/Source/TypeHandlers/Groups/DisplayGroupBBox.cpp
@@ -14,8 +14,8 @@ DisplayGroupBBox::DisplayGroupBBox()
 {
 }
 
-void	DisplayGroupBBox::Process( uint32_t inStartOffset,
-									uint32_t inEndOffset )
+void	DisplayGroupBBox::Process( size_t inStartOffset,
+									size_t inEndOffset )
 {
 	if (inEndOffset != inStartOffset + 24)
 	{
