+++ SDK/Extras/Textify3DMF/Source/TypeHandlers/Lights/PointLight.cpp
@@ -32,8 +32,8 @@ PointLight::PointLight()
 {
 }
 
-void	PointLight::Process( uint32_t inStartOffset,
-									uint32_t inEndOffset )
+void	PointLight::Process( size_t inStartOffset,
+									size_t inEndOffset )
 {
 	if (inEndOffset != inStartOffset + 28)
 	{
