+++ SDK/Extras/Textify3DMF/Source/TypeHandlers/Miscellaneous/TableOfContents.h
@@ -31,9 +31,9 @@ class TableOfContents : public TypeHandler
 public:
 							TableOfContents();
 	
-	void					PreParse( uint32_t inStartOffset,
-									uint32_t inEndOffset );
+	void					PreParse( size_t inStartOffset,
+									size_t inEndOffset );
 	
-	virtual void			Process( uint32_t inStartOffset,
-									uint32_t inEndOffset );
+	virtual void			Process( size_t inStartOffset,
+									size_t inEndOffset );
 };
