+++ SDK/Extras/Textify3DMF/Source/TypeHandlers/Styles/AntialiasStyle.cpp
@@ -32,8 +32,8 @@ AntialiasStyle::AntialiasStyle()
 {
 }
 
-void	AntialiasStyle::Process( uint32_t inStartOffset,
-									uint32_t inEndOffset )
+void	AntialiasStyle::Process( size_t inStartOffset,
+									size_t inEndOffset )
 {
 	if (inEndOffset != inStartOffset + 12)
 	{
