+++ SDK/Extras/Textify3DMF/Source/TypeHandlers/Styles/BackfacingStyle.cpp
@@ -43,8 +43,8 @@ BackfacingStyle::BackfacingStyle()
 {
 }
 
-void	BackfacingStyle::Process( uint32_t inStartOffset,
-									uint32_t inEndOffset )
+void	BackfacingStyle::Process( size_t inStartOffset,
+									size_t inEndOffset )
 {
 	if (inEndOffset != inStartOffset + 4)
 	{
