+++ SDK/Extras/Textify3DMF/Source/TypeHandlers/Styles/CastShadowsStyle.cpp
@@ -32,8 +32,8 @@ CastShadowsStyle::CastShadowsStyle()
 {
 }
 
-void	CastShadowsStyle::Process( uint32_t inStartOffset,
-									uint32_t inEndOffset )
+void	CastShadowsStyle::Process( size_t inStartOffset,
+									size_t inEndOffset )
 {
 	if (inEndOffset != inStartOffset + 4)
 	{
