+++ SDK/Extras/Textify3DMF/Source/TypeHandlers/Styles/FillStyle.cpp
@@ -42,8 +42,8 @@ FillStyle::FillStyle()
 {
 }
 
-void	FillStyle::Process( uint32_t inStartOffset,
-									uint32_t inEndOffset )
+void	FillStyle::Process( size_t inStartOffset,
+									size_t inEndOffset )
 {
 	if (inEndOffset != inStartOffset + 4)
 	{
