+++ SDK/Extras/Textify3DMF/Source/TypeHandlers/Styles/LineWidthStyle.cpp
@@ -32,8 +32,8 @@ LineWidthStyle::LineWidthStyle()
 {
 }
 
-void	LineWidthStyle::Process( uint32_t inStartOffset,
-									uint32_t inEndOffset )
+void	LineWidthStyle::Process( size_t inStartOffset,
+									size_t inEndOffset )
 {
 	if (inEndOffset != inStartOffset + 4)
 	{
