+++ SDK/Extras/Textify3DMF/Source/TypeHandlers/Styles/PickIDStyle.cpp
@@ -32,8 +32,8 @@ PickIDStyle::PickIDStyle()
 {
 }
 
-void	PickIDStyle::Process( uint32_t inStartOffset,
-									uint32_t inEndOffset )
+void	PickIDStyle::Process( size_t inStartOffset,
+									size_t inEndOffset )
 {
 	if (inEndOffset != inStartOffset + 4)
 	{
