+++ SDK/Extras/Textify3DMF/Source/TypeHandlers/Transforms/RotateAboutAxis.cpp
@@ -32,8 +32,8 @@ RotateAboutAxis::RotateAboutAxis()
 {
 }
 
-void	RotateAboutAxis::Process( uint32_t inStartOffset,
-										uint32_t inEndOffset )
+void	RotateAboutAxis::Process( size_t inStartOffset,
+										size_t inEndOffset )
 {
 	if (inEndOffset != inStartOffset + 28)
 	{
